[BE/ENV] 음식 정보 DB에 입력

See merge request s09-bigdata-recom-sub2/S09P22A403!38

+++ backend/fns/src/main/java/ssafy/fns/domain/exercise/entity/Exercise.java
@@ -38,12 +38,12 @@ public class Exercise extends BaseEntity {
 
     private LocalDateTime date;
 
-    private Long kcal;
+    private Double kcal;
 
     private Long exerciseTime;
 
     @Builder
-    public Exercise(Sports sports, Member member, LocalDateTime date, Long kcal,Long exerciseTime) {
+    public Exercise(Sports sports, Member member, LocalDateTime date, Double kcal,Long exerciseTime) {
         this.sports = sports;
         this.member = member;
         this.date = date;
